Added Node class, updated args and created a base main.py
